+++ packages/at_client/lib/src/service/sync_service_impl.dart
@@ -166,7 +166,8 @@ class SyncServiceImpl implements SyncService {
         atClient.getCurrentAtSign()!, atClient.getPreferences()!,
         atChops: atClient.atChops,
         enrollmentId: atClient.enrollmentId,
-        signingAlgoType: signingAlgoOf(atClient));
+        signingAlgoType: signingAlgoOf(atClient),
+        atKeysIo: atClient.atKeysIo);
     final syncService = SyncServiceImpl._(atClient, remoteSecondary);
     await syncService.statsServiceListener();
     syncService._startPeriodicSyncTimer();